+++ tests/functional/test_run_synthesis.py
@@ -45,10 +45,10 @@ def test_synthesis(tmp_path):
         ).resolve()
         assert edges_file.stat().st_size > 100
         # test output from choose_morphologies
-        assert Path("axon-morphologies.tsv").stat().st_size > 100
+        assert Path("auxiliary/axon-morphologies.tsv").stat().st_size > 100
         # test output from synthesize_morphologies
-        assert Path("circuit.synthesized_morphologies.h5").stat().st_size > 100
-        assert Path("circuit.ais_scales.h5").stat().st_size > 100
+        assert Path("auxiliary/circuit.synthesized_morphologies.h5").stat().st_size > 100
+        assert Path("auxiliary/circuit.ais_scales.h5").stat().st_size > 100
 
         with h5py.File(edges_file, "r") as h5f:
             assert f"/edges/{edge_population_name}" in h5f
@@ -64,9 +64,9 @@ def test_synthesis(tmp_path):
             config = json.load(f)
             assert (
                 config["networks"]["nodes"][0]["nodes_file"]
-                == f"$NETWORK_NODES_DIR/{node_population_name}/nodes.h5"
+                == f"$BASE_DIR/networks/nodes/{node_population_name}/nodes.h5"
             )
             assert (
                 config["networks"]["edges"][0]["edges_file"]
-                == f"$NETWORK_EDGES_DIR/{edge_population_name}/edges.h5"
+                == f"$BASE_DIR/networks/edges/functional/{edge_population_name}/edges.h5"
             )